// generated/addon/angular/src/app/api-models/cieb-external-carrier.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebExternalCarrier {
  seqExtnId: number;
  seqPremId: number;
  extnCarrGroupId: string;
  productType: string;
  extnCarrName: string;
  extnCarrId: string;
  sharMethod: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  sharBenType: string;
}